# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AIGYPT Poster Challenge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Create or replace a submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created/replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Validation error or deadline passed
  /submissions/by-participant/{participantId}:
    get:
      operationId: getSubmissionByParticipant
      tags: [submissions]
      summary: Get submission for a participant
      parameters:
        - name: participantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submission found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: No submission yet
...
  /votes:
    post:
      operationId: castVote
      tags: [votes]
      summary: Cast a vote for a submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "201":
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "400":
          description: Voting closed or limit reached
  /votes/{submissionId}:
    delete:
      operationId: removeVote
      tags: [votes]
      summary: Remove a vote from a submission
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vote removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
  /prize/claim:
    post:
      operationId: claimPrize
      tags: [prizes]
      summary: Claim prize code (server-side, service role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimPrizeInput"
      responses:
        "200":
          description: Prize code(s) returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimedPrize"
        "400":
          description: Not eligible or already claimed
...
components:
  schemas:
...
    SubmissionInput:
      type: object
      required: [participant_id, image_url, ig_post_url]
      properties:
        participant_id:
          type: string
        image_url:
          type: string
        ig_post_url:
          type: string
    Submission:
      type: object
      required: [id, participant_id, image_url, ig_post_url, status, created_at]
      properties:
        id:
          type: string
        participant_id:
          type: string
        image_url:
          type: string
        ig_post_url:
          type: string
        status:
          type: string
        winner_category:
          type: ["string", "null"]
        created_at:
          type: string
...
    VoteInput:
      type: object
      required: [submission_id, voter_fingerprint]
      properties:
        submission_id:
          type: string
        voter_fingerprint:
          type: string
    VoteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        vote_count:
          type: integer
    ClaimPrizeInput:
      type: object
      required: [participant_id]
      properties:
        participant_id:
          type: string
    ClaimedPrize:
      type: object
      required: [code, tier]
      properties:
        code:
          type: string
        tier:
          type: string
...