# lib/api-zod/src/generated/types/submissionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AIGYPT Poster Challenge API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionInput {
  participant_id: string;
  image_url: string;
  ig_post_url: string;
}