# lib/api-zod/src/generated/types/voteInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AIGYPT Poster Challenge API
 * OpenAPI spec version: 0.1.0
 */

export interface VoteInput {
  submission_id: string;
  voter_fingerprint: string;
}