# lib/api-zod/src/generated/types/voteResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AIGYPT Poster Challenge API
 * OpenAPI spec version: 0.1.0
 */

export interface VoteResult {
  success: boolean;
  message?: string;
  vote_count?: number;
}